+++ cmd/provenanced/cmd/root.go
@@ -157,6 +157,13 @@ func initRootCmd(rootCmd *cobra.Command, encodingConfig params.EncodingConfig) {
 
 	// Add Rosetta command
 	rootCmd.AddCommand(server.RosettaCommand(encodingConfig.InterfaceRegistry, encodingConfig.Marshaler))
+
+	// Disable usage when the start command returns an error.
+	startCmd, _, err := rootCmd.Find([]string{"start"})
+	if err != nil {
+		panic(fmt.Errorf("start command not found: %w", err))
+	}
+	startCmd.SilenceUsage = true
 }
 
 func addModuleInitFlags(startCmd *cobra.Command) {